requirements.txt
torch==2.5.0
pyyaml==6.0.3
imageio==2.37.2
fastapi==0.115.12
uvicorn==0.35.0
pydantic==2.12.5
supabase==2.27.2
python-dotenv==1.0.0
pika==1.3.2
cloudinary==1.44.1
Pillow==10.4.0
numpy==1.26.4
torchvision==0.20.0
